feat: Add new test case and update vehicle id to long type

+++ CarSales.Web.Domain/Interfaces/IVehicleRepository.cs
@@ -5,7 +5,7 @@ namespace CarSales.Web.Domain.Interfaces
 {
     public interface IVehicleRepository
     {
-        public Task<Vehicle> GetVehicleAsync(int vehicleId);
+        public Task<Vehicle> GetVehicleAsync(long vehicleId);
 
         public Task AddVehicleAsync(Vehicle vehicle);
